pyproject.toml
[tool.poetry]
name = "ghost-gpt"
version = "0.1.0"
description = ""
authors = ["Bram Adams <3282661+bramses@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ghost_gpt"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
openai = "^1.12.0"
scikit-learn = "^1.4.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.6.0"
pydantic = "^2.6.1"
uvicorn = "^0.27.1"
jinja2 = "^3.1.3"
pytailwindcss = "^0.2.0"
supabase = "^2.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:start"